+++ internal/otelcollector/config/log/gateway/processors.go
@@ -2,7 +2,7 @@ package gateway
 
 import (
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/gatewayprocs"
+	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/processors"
 )
 
 func makeProcessorsConfig(opts BuildOptions) Processors {
@@ -11,8 +11,8 @@ func makeProcessorsConfig(opts BuildOptions) Processors {
 			Batch:         makeBatchProcessorConfig(),
 			MemoryLimiter: makeMemoryLimiterConfig(),
 		},
-		K8sAttributes:           gatewayprocs.K8sAttributesProcessorConfig(opts.Enrichments),
-		InsertClusterAttributes: gatewayprocs.InsertClusterAttributesProcessorConfig(opts.ClusterName, opts.CloudProvider),
+		K8sAttributes:           processors.K8sAttributesProcessorConfig(opts.Enrichments),
+		InsertClusterAttributes: processors.InsertClusterAttributesProcessorConfig(opts.ClusterName, opts.CloudProvider),
 	}
 }
