+++ internal/otelcollector/config/processors/k8s_attribute_proc_test.go
@@ -1,4 +1,4 @@
-package gatewayprocs
+package processors
 
 import (
 	"testing"
