+++ internal/otelcollector/config/processors/resolve_service_name_proc.go
@@ -1,4 +1,4 @@
-package gatewayprocs
+package processors
 
 import (
 	"fmt"
