+++ internal/otelcollector/config/processors/resolve_service_name_proc_test.go
@@ -1,4 +1,4 @@
-package gatewayprocs
+package processors
 
 import (
 	"testing"
