+++ internal/otelcollector/config/processors/resource_procs.go
@@ -1,4 +1,4 @@
-package gatewayprocs
+package processors
 
 import (
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
