+++ internal/otelcollector/config/processors/resource_procs_test.go
@@ -1,4 +1,4 @@
-package gatewayprocs
+package processors
 
 import (
 	"testing"
